Pulling in some history changes from the original myvideobuzz developers
Video history is limited to the last 5 videos.
It can be cleared by going to the Settings page and hitting 'Clear
History'
The LAN videos channel now uses the first item in the history list to
respond to queries.
The history appears to use roughly 4kb of registry space when full of 5
items.

This addresses Issue #28