Adds 'View Full Description" menu option (* menu) this shows the title and full description in a text screen
Yes, I know the text is small, but my options are limited. This
addresses Issue #54.
You can also use the * menu on the video details screen now to change
the sleep timer and to 'View Full Description'
Fixed potential crash with 'Play All' if it tries to 'Play' the next
arrow
Fixed using left and right arrows on the video details screen going to
the previous/next buttons.