Fixes the annoyance on screens with the filter banner
When you hit 'Play' on the screen with categories
(Subscriptions/Playlists/Reddit/etc) and the video finishes playing the
category would be selected. The video that finished playing should now
be selected instead.

Also, if you change the video selection in the Details view (either via
Play All, or the left/right buttons) and then return to the video list,
the video you had been viewing in the details will now be selected.